+++ src/main/java/dev/gacbl/logicore/LogiCore.java
@@ -1,6 +1,8 @@
 package dev.gacbl.logicore;
 
 import com.mojang.logging.LogUtils;
+import dev.gacbl.logicore.blocks.battery.BatteryFillRenderer;
+import dev.gacbl.logicore.blocks.battery.advance.AdvanceBatteryModule;
 import dev.gacbl.logicore.blocks.battery.basic.BasicBatteryModule;
 import dev.gacbl.logicore.blocks.cloud_interface.CloudInterfaceModule;
 import dev.gacbl.logicore.blocks.compiler.CompilerBlockEntityRenderer;
@@ -72,6 +74,7 @@ public class LogiCore {
         WrenchModule.register(modEventBus);
         DroneModule.register(modEventBus);
         BasicBatteryModule.register(modEventBus);
+        AdvanceBatteryModule.register(modEventBus);
         IntegrationUtils.registerEvents();
 
         PacketHandler.register(modEventBus);
@@ -125,6 +128,7 @@ public class LogiCore {
         @SubscribeEvent
         public static void registerRenderers(EntityRenderersEvent.RegisterRenderers event) {
             event.registerBlockEntityRenderer(DroneBayModule.DRONE_BAY_BE.get(), DroneBayRenderer::new);
+            event.registerBlockEntityRenderer(AdvanceBatteryModule.ADVANCE_BATTERY_BE.get(), BatteryFillRenderer::new);
         }
 
         @SubscribeEvent