+++ CHANGELOG.md
@@ -25,5 +25,7 @@
     - It allows storing cycles in the "cloud" and sharing it with the team.
     - It can be used to transfer cycles from the cloud to machines as well
 - Added wrench item
-- Added basic battery block for energy storage
+- Added battery blocks for energy storage
+    - Basic
+    - Advance
 - Added AE2 integration via Cloud Interface block